requirements.txt
amqp==2.5.2
attrs==19.3.0
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9
cssselect==1.1.0
Django==2.1.8
django-cors-headers==3.2.1
django-haystack==2.8.1
django-ranged-response==0.2.0
django-redis==4.11.0
django-simple-captcha==0.5.12
elasticsearch==2.4.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
kombu==4.6.8
lxml==4.5.0
parsel==1.5.2
Pillow==7.1.1
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pytz==2019.3
queuelib==1.5.0
redis==3.4.1
requests==2.23.0
Scrapy==2.0.1
service-identity==18.1.0
six==1.14.0
Twisted==20.3.0
urllib3==1.25.9
vine==1.3.0
w3lib==1.21.0
zipp==3.1.0
zope.interface==5.1.0
